+++ cli/src/lib/storage-changes/memory-snapshot.ts
